+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Added
-
 - Listring add for circular saw.
+- The circular saw can no longer replace items from the player's inventory
+  when it is full.
 
 ## [1.1.0] - 2017-10-04
 
@@ -35,5 +35,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial versioned release.
 
+[Unreleased]: https://github.com/minetest-mods/moreblocks/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/minetest-mods/moreblocks/compare/v1.0.0...v1.1.0
-[Unreleased]: https://github.com/minetest-mods/moreblocks/compare/v1.0.0...HEAD